+++ R/DesignClass.R
@@ -28,7 +28,8 @@ Design <- setClass(Class = "Design",
                              content = 'factor',
                              content_prop = 'numeric',
                              content_prop_empirical = 'numeric',
-                             constraints = 'list'),
+                             constraints = 'list',
+                             customNextItem = 'function'),
                    validity = function(object) return(TRUE)
 )
 
@@ -78,7 +79,8 @@ setMethod("initialize", signature(.Object = "Design"),
                   dnames <- names(design)
                   gnames <- c('min_SEM', 'thetas.start', 'min_items', 'max_items', 'quadpts', 
                               'theta_range', 'weights', 'KL_delta', 'content', 'content_prop',
-                              'classify', 'classify_CI', 'exposure', 'delta_thetas', 'constraints')
+                              'classify', 'classify_CI', 'exposure', 'delta_thetas', 'constraints',
+                              'customNextItem')
                   if(!all(dnames %in% gnames))
                       stop('The following inputs to design are invalid: ',
                            paste0(dnames[!(dnames %in% gnames)], ' '), call.=FALSE)
@@ -149,6 +151,10 @@ setMethod("initialize", signature(.Object = "Design"),
                         stop('The first item can not be used in an \'independent\' constraint. 
                               Consider removing the items that will not be used from the test.', 
                              call.=FALSE)
+                  }
+                  if(!is.null(design$customNextItem)){
+                      .Object@customNextItem <- design$customNextItem
+                      .Object@criteria <- 'custom'
                   }
               }
               if(.Object@use_content && criteria == 'seq')