+++ tests/tests/test-mirtCAT_uni.R
@@ -47,6 +47,15 @@ test_that('unidimensional', {
     res <- mirtCAT(df, local_pattern=pat, criteria='random')
     expect_true(all(!is.na(res$raw_responses)))
     
+    # custom
+    customNextItem <- function(person, design, test, thetas){
+        sum(is.na(person$items_answered)) + 1L
+    }
+    res <- mirtCAT(df, local_pattern=pat, design = list(customNextItem=customNextItem))
+    expect_is(res, 'mirtCAT')
+    so <- summary(res)
+    expect_equal(c(1, 25:2), so$items_answered)
+    
     #sequential
     res <- mirtCAT(df2, mod, local_pattern=pat)
     expect_equal(as.numeric(res$thetas), 0.3588322, tolerance = 1e-4)