+++ R/createShinyGUI.R
@@ -5,6 +5,9 @@
 #' through \url{http://www.shinyapps.io/}. The function \code{\link{mirtCAT_preamble}} must be 
 #' run \emph{before} this function is called. The object is executed by calling \code{\link{runApp}}.
 #' 
+#' @param ui a shiny UI function used to define the interface. If \code{NULL}, the default one will be used. 
+#'   See \code{mirtCAT:::default_UI} for the internal code
+#' 
 #' @export createShinyGUI
 #' 
 #' @author Phil Chalmers \email{rphilip.chalmers@@gmail.com}
@@ -21,9 +24,10 @@
 #' summary(person)
 #' 
 #' } 
-createShinyGUI <- function(){
+createShinyGUI <- function(ui = NULL){
     on.exit(.MCE$preamble_defined <- NULL)
     if(is.null(.MCE$preamble_defined))
         stop('Please use a fresh mirtCAT_preamble() call prior to calling createShinyGUI().')
+    if(is.null(ui)) ui <- default_UI
     return(shinyApp(ui=ui(), server=server))
 }